requirements.txt
absl-py==2.1.0
amqp==5.3.1
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.4.0
arrow==1.3.0
asgiref==3.7.2
assemblyai==0.30.0
asttokens==2.4.1
astunparse==1.6.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
beautifulsoup4==4.12.3
billiard==4.2.1
boto3==1.35.32
botocore==1.35.32
cachetools==5.5.0
celery==5.4.0
certifi==2023.11.17
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cmudict==1.0.31
colorama==0.4.6
constantly==23.10.4
contourpy==1.2.1
coverage==7.6.4
coveralls==4.0.1
cryptography==41.0.7
cycler==0.12.1
daphne==4.1.2
decorator==5.1.1
defusedxml==0.7.1
Distance==0.1.3
distro==1.9.0
Django==5.0.1
django-allauth==0.60.1
django-appconf==1.0.6
django-cors-headers==4.4.0
django-crum==0.7.9
django-extensions==3.2.3
django-filter==24.3
django-image-cropping==1.7
django-jazzmin==3.0.0
django-modelcluster==6.3
django-permissionedforms==0.1
django-phonenumber-field==8.0.0
django-ses==4.2.0
django-storages==1.14.4
django-taggit==6.1.0
django-treebeard==4.7.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
draftjs-exporter==5.0.0
drf-rw-serializers==1.4.0
drf-spectacular==0.27.2
easy-thumbnails==2.8.5
elevenlabs==0.3.0b0
et_xmlfile==2.0.0
executing==2.0.1
fake-bpy-module==20240712
ffmpeg-python==0.2.0
filelock==3.16.1
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.53.0
fpdf==1.7.2
fsspec==2024.9.0
future==1.0.0
g2p-en==2.1.0
gast==0.5.4
git-filter-repo==2.45.0
google-ai-generativelanguage==0.6.10
google-api-core==2.20.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-speech==2.27.0
google-cloud-texttospeech==2.18.0
google-generativeai==0.8.3
google-pasta==0.2.0
googleapis-common-protos==1.65.0
grpcio==1.67.0
grpcio-status==1.66.2
gTTS==2.5.3
h11==0.14.0
h2==4.1.0
h5py==3.11.0
hpack==4.0.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.25.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.6
imageio-ffmpeg==0.5.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
incremental==24.7.2
inflect==7.4.0
inflection==0.5.1
ipython==8.26.0
jedi==0.19.1
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.3.3
kiwisolver==1.4.5
kombu==5.4.2
l18n==2021.3
laces==0.1.1
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.3.2
more-itertools==10.5.0
namex==0.0.8
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.35.14
openpyxl==3.1.5
opt-einsum==3.3.0
optree==0.11.0
packaging==24.1
pandas==2.2.1
parso==0.8.4
phonenumbers==8.13.49
pillow==10.2.0
pillow_heif==0.20.0
pocketsphinx==5.0.3
priority==1.3.0
prompt-toolkit==3.0.43
pronouncing==0.2.0
proto-plus==1.24.0
protobuf==5.28.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
pure-eval==0.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydub==0.25.1
pygame==2.6.1
Pygments==2.18.0
PyJWT==2.8.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.2
razorpay==1.4.1
redis==5.2.0
referencing==0.35.1
regex==2024.9.11
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.1
rpds-py==0.21.0
rsa==4.9
s3transfer==0.10.2
safetensors==0.4.5
scikit-learn==1.5.0
scipy==1.13.1
service-identity==24.1.0
setuptools==69.1.1
six==1.16.0
sniffio==1.3.1
sounddevice==0.5.0
soupsieve==2.5
SpeechRecognition==3.10.4
sqlparse==0.4.4
srt==3.5.3
stack-data==0.6.3
telepath==0.3.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
termcolor==2.4.0
threadpoolctl==3.5.0
tokenizers==0.19.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.44.2
Twisted==24.7.0
txaio==23.1.1
typeguard==4.3.0
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
vosk==0.3.45
wagtail==6.3
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.3
wheel==0.43.0
whitenoise==6.7.0
Willow==1.9.0
wit==6.0.1
wrapt==1.16.0
XlsxWriter==3.2.0
zipp==3.20.2
zope.interface==7.0.3
